requirements.txt
flask
pygpt4all